1) Refactor code structure. 2) Fix lint errors. 3) Remove demo test cases (To be re-added as future unit tests).
