+++ knowhere/knowhere/index/vector_index/impl/bruteforce/BruteForce.cpp
@@ -1,5 +1,15 @@
-#include "BruteForce.h"
+// Copyright (C) 2019-2020 Zilliz. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
+// with the License. You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under the License
+// is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+// or implied. See the License for the specific language governing permissions and limitations under the License
 #include <omp.h>
+#include "knowhere/index/vector_index/impl/bruteforce/BruteForce.h"
 
 namespace knowhere {
 
@@ -65,15 +75,15 @@ void knn_inner_product_sse(const float * x,
 
         for (size_t j = 0; j < ny; j++) {
             if (!bitset || !bitset.test(j)) {
-                float disij = faiss::fvec_inner_product_ref (x_i, y_j, d);
+                float disij = faiss::fvec_inner_product_ref(x_i, y_j, d);
                 if (disij > val_[0]) {
-                    minheap_swap_top (k, val_, ids_, disij, j);
+                    minheap_swap_top(k, val_, ids_, disij, j);
                 }
             }
             y_j += d;
         }
 
-        minheap_reorder (k, val_, ids_);
+        minheap_reorder(k, val_, ids_);
     }
 }
