+++ knowhere/knowhere/index/vector_index/impl/bruteforce/SimpleIndexFlat.h
@@ -1,3 +1,14 @@
+// Copyright (C) 2019-2020 Zilliz. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
+// with the License. You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under the License
+// is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+// or implied. See the License for the specific language governing permissions and limitations under the License
+
 #pragma once
 /**
  *  Simplified version of Faiss's IndexFlat struct with minimal functions.
@@ -66,7 +77,6 @@ struct SimpleIndexFlat {
             const faiss::BitsetView bitset = nullptr) const;
 
     void train(idx_t n, const float* x);
-
 };
 
 }  // namespace knowhere
