--- knowhere/knowhere/index/vector_index/impl/bruteforce/test/test_demo_sift1M.cpp
@@ -1,386 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-#include <cstdio>
-#include <cstdlib>
-#include <cassert>
-#include <cstring>
-
-#include <algorithm>
-#include <limits>
-
-#include <sys/stat.h>
-
-#include <sys/time.h>
-
-#include "knowhere/index/vector_index/impl/bruteforce/indexing/SimpleIndexFlat.h"
-
-/**
- * To run this demo, please download the ANN_SIFT1M dataset from
- *
- *   http://corpus-texmex.irisa.fr/
- *
- * and unzip it to the sudirectory sift1M.
- **/
-
-/*****************************************************
- * I/O functions for fvecs and ivecs
- *****************************************************/
-
-
-float * fvecs_read (const char *fname,
-                    size_t *d_out, size_t *n_out)
-{
-    FILE *f = fopen(fname, "r");
-    if(!f) {
-        fprintf(stderr, "could not open %s\n", fname);
-        perror("");
-        abort();
-    }
-    int d;
-    fread(&d, 1, sizeof(int), f);
-    assert((d > 0 && d < 1000000) || !"unreasonable dimension");
-    fseek(f, 0, SEEK_SET);
-    struct stat st;
-    fstat(fileno(f), &st);
-    size_t sz = st.st_size;
-    assert(sz % ((d + 1) * 4) == 0 || !"weird file size");
-    size_t n = sz / ((d + 1) * 4);
-
-    *d_out = d; *n_out = n;
-    float *x = new float[n * (d + 1)];
-    size_t nr = fread(x, sizeof(float), n * (d + 1), f);
-    assert(nr == n * (d + 1) || !"could not read whole file");
-
-    // shift array to remove row headers
-    for(size_t i = 0; i < n; i++)
-        memmove(x + i * d, x + 1 + i * (d + 1), d * sizeof(*x));
-
-    fclose(f);
-    return x;
-}
-
-// not very clean, but works as long as sizeof(int) == sizeof(float)
-int *ivecs_read(const char *fname, size_t *d_out, size_t *n_out)
-{
-    return (int*)fvecs_read(fname, d_out, n_out);
-}
-
-double elapsed ()
-{
-    struct timeval tv;
-    gettimeofday (&tv, nullptr);
-    return  tv.tv_sec + tv.tv_usec * 1e-6;
-}
-
-float CalcRecall(int64_t topk, int64_t k, int nq, knowhere::SimpleIndexFlat::idx_t* gt, knowhere::SimpleIndexFlat::idx_t* nt) {
-    float sum_ratio = 0.0f;
-    for (int i = 0; i < nq; i++) {
-        //std::vector<int64_t> ids_0 = true_ids[i].ids;
-        //std::vector<int64_t> ids_1 = result_ids[i].ids;
-        std::vector<knowhere::SimpleIndexFlat::idx_t> ids_0(gt+i*k,gt+i*k+topk);
-        std::vector<knowhere::SimpleIndexFlat::idx_t> ids_1(nt+i*k,nt+i*k+topk);
-        std::sort(ids_0.begin(), ids_0.end());
-        std::sort(ids_1.begin(), ids_1.end());
-        std::vector<knowhere::SimpleIndexFlat::idx_t> v(nq * 2);
-        std::vector<knowhere::SimpleIndexFlat::idx_t>::iterator it;
-        it=std::set_intersection (ids_0.begin(), ids_0.end(), ids_1.begin(), ids_1.end(), v.begin());
-        v.resize(it-v.begin());
-        sum_ratio += 1.0f * v.size() / topk;
-    }
-    return 1.0 * sum_ratio / nq;
-}
-
-int main() {
-    size_t d;
-    size_t nb;
-
-    size_t loops = 5;
-
-    float *xb = fvecs_read("sift1M/sift_base.fvecs", &d, &nb);
-
-    int nlist = 100;
-    int m = 8;
-    // knowhere::SimpleIndexFlatFlatL2 quantizer(d);
-    // knowhere::SimpleIndexFlatIVFPQ index(&quantizer, d, nlist, m, 8);
-
-    knowhere::SimpleIndexFlat index(d, knowhere::METRIC_L2);
-    //index.setNumProbes(10);
-
-    index.train(nb, xb);
-    printf("is_trained = %s\n", index.is_trained ? "true" : "false");
-    index.add(nb, xb);                     // add vectors to the index
-    printf("ntotal = %lld\n", index.ntotal);
-
-    size_t nq;
-    float *xq;
-
-    size_t d2;
-    xq = fvecs_read("sift1M/sift_query.fvecs", &d2, &nq);
-    assert(d == d2 || !"query does not have same dimension as train set");
-    size_t k; // nb of results per query in the GT
-    knowhere::SimpleIndexFlat::idx_t *gt;  // nq * k matrix of ground-truth nearest-neighbors
-
-    // load ground-truth and convert int to long
-    size_t nq2;
-    int *gt_int = ivecs_read("sift1M/sift_groundtruth.ivecs", &k, &nq2);
-    assert(nq2 == nq || !"incorrect nb of ground truth entries");
-
-    printf("d: %ld\n", d);
-    printf("nb: %ld\n", nb);
-    printf("nq: %ld\n", nq);
-    printf("k: %ld\n", k);
-
-    gt = new knowhere::SimpleIndexFlat::idx_t[k * nq];
-    for(int i = 0; i < k * nq; i++) {
-        gt[i] = gt_int[i];
-    }
-    delete [] gt_int;
-
-    {
-        long long *I = new long long[k * nq];
-        float *D = new float[k * nq];
-        index.search(nq, xq, k, D, I);
-        double avg = 0.0f, dur;
-        double min_time = std::numeric_limits<double>::max();
-
-        for (int i = 0; i < loops; i++) {
-            double t0 = elapsed();
-            index.search(nq, xq, k, D, I);
-
-            dur = elapsed() - t0;
-            avg += dur;
-            min_time = std::min(min_time, dur);
-        }
-        avg /= loops;
-        printf("Recall: %.4f\nTime spent: %.3fs\nMin time: %.3fs\n", CalcRecall(k, k, nq, gt, I), avg, min_time);
-        delete [] I;
-        delete [] D;
-    }
-
-
-#if 0
-    // First SQ8, then IndexRefineFlat
-    {
-        long *I = new long[k * nq];
-        float *D = new float[k * nq];
-        knowhere::SimpleIndexFlat* sq = new knowhere::SimpleIndexFlatScalarQuantizer(d, knowhere::QuantizerType::QT_8bit);
-        knowhere::SimpleIndexFlatRefineFlat rf = knowhere::SimpleIndexFlatRefineFlat(sq);
-        rf.train(nb, xb);
-        rf.add(nb, xb);
-        rf.k_factor = 4;
-        rf.search(nq, xq, k, D, I);
-        double avg = 0.0f;
-        for (int i = 0; i < loops; i++) {
-            double t0 = elapsed();
-            rf.search(nq, xq, k, D, I);
-            avg += elapsed() - t0;
-        }
-        avg /= loops;
-        printf("RefineFlat Recall: %.4f, time spent: %.3fs\n", CalcRecall(k, k, nq, gt, I), avg);
-        delete [] I;
-        delete [] D;
-    }
-
-    // SQ8 alone
-    {
-        long *I = new long[k * nq];
-        float *D = new float[k * nq];
-        auto sq = new knowhere::SimpleIndexFlatScalarQuantizer(d, knowhere::QuantizerType::QT_8bit);
-        sq->train(nb, xb);
-        sq->add(nb, xb);
-        sq->search(nq, xq, k, D, I);
-        double avg = 0.0f;
-        for (int i = 0; i < loops; i++) {
-            double t0 = elapsed();
-            sq->search(nq, xq, k, D, I);
-            avg += elapsed() - t0;
-        }
-        avg /= loops;
-        printf("SQ8 Recall: %.4f, time spent: %.3fs\n", CalcRecall(k, k, nq, gt, I), avg);
-        delete [] I;
-        delete [] D;
-    }
-#endif
-
-    delete [] xq;
-    delete [] gt;
-    return 0;
-}
-
-#if 0
-int main()
-{
-    double t0 = elapsed();
-
-    // this is typically the fastest one.
-    const char *index_key = "IVF2048,Flat";
-
-    // these ones have better memory usage
-    // const char *index_key = "Flat";
-    // const char *index_key = "PQ32";
-    // const char *index_key = "PCA80,Flat";
-    // const char *index_key = "IVF4096,PQ8+16";
-    // const char *index_key = "IVF4096,PQ32";
-    // const char *index_key = "IMI2x8,PQ32";
-    // const char *index_key = "IMI2x8,PQ8+16";
-    // const char *index_key = "OPQ16_64,IMI2x8,PQ8+16";
-
-    knowhere::SimpleIndexFlat * index;
-
-    size_t d;
-
-    {
-        printf ("[%.3f s] Loading train set\n", elapsed() - t0);
-
-        size_t nt;
-        float *xt = fvecs_read("sift1M/sift_learn.fvecs", &d, &nt);
-
-        printf ("[%.3f s] Preparing index \"%s\" d=%ld\n",
-                elapsed() - t0, index_key, d);
-        index = knowhere::SimpleIndexFlat_factory(d, index_key);
-
-        printf ("[%.3f s] Training on %ld vectors\n", elapsed() - t0, nt);
-
-        index->train(nt, xt);
-        delete [] xt;
-    }
-
-
-    {
-        printf ("[%.3f s] Loading database\n", elapsed() - t0);
-
-        size_t nb, d2;
-        float *xb = fvecs_read("sift1M/sift_base.fvecs", &d2, &nb);
-        assert(d == d2 || !"dataset does not have same dimension as train set");
-
-        printf ("[%.3f s] Indexing database, size %ld*%ld\n",
-                elapsed() - t0, nb, d);
-
-        index->add(nb, xb);
-
-        delete [] xb;
-    }
-
-    size_t nq;
-    float *xq;
-
-    {
-        printf ("[%.3f s] Loading queries\n", elapsed() - t0);
-
-        size_t d2;
-        xq = fvecs_read("sift1M/sift_query.fvecs", &d2, &nq);
-        assert(d == d2 || !"query does not have same dimension as train set");
-
-    }
-
-    size_t k; // nb of results per query in the GT
-    knowhere::SimpleIndexFlat::idx_t *gt;  // nq * k matrix of ground-truth nearest-neighbors
-
-    {
-        printf ("[%.3f s] Loading ground truth for %ld queries\n",
-                elapsed() - t0, nq);
-
-        // load ground-truth and convert int to long
-        size_t nq2;
-        int *gt_int = ivecs_read("sift1M/sift_groundtruth.ivecs", &k, &nq2);
-        assert(nq2 == nq || !"incorrect nb of ground truth entries");
-
-        gt = new knowhere::SimpleIndexFlat::idx_t[k * nq];
-        for(int i = 0; i < k * nq; i++) {
-            gt[i] = gt_int[i];
-        }
-        delete [] gt_int;
-    }
-
-    // // Result of the auto-tuning
-    // std::string selected_params;
-
-    // { // run auto-tuning
-
-    //     printf ("[%.3f s] Preparing auto-tune criterion 1-recall at 1 "
-    //             "criterion, with k=%ld nq=%ld\n", elapsed() - t0, k, nq);
-
-    //     knowhere::OneRecallAtRCriterion crit(nq, 1);
-    //     crit.set_groundtruth (k, nullptr, gt);
-    //     crit.nnn = k; // by default, the criterion will request only 1 NN
-
-    //     printf ("[%.3f s] Preparing auto-tune parameters\n", elapsed() - t0);
-
-    //     knowhere::ParameterSpace params;
-    //     params.initialize(index);
-
-    //     printf ("[%.3f s] Auto-tuning over %ld parameters (%ld combinations)\n",
-    //             elapsed() - t0, params.parameter_ranges.size(),
-    //             params.n_combinations());
-
-    //     knowhere::OperatingPoints ops;
-    //     params.explore (index, nq, xq, crit, &ops);
-
-    //     printf ("[%.3f s] Found the following operating points: \n",
-    //             elapsed() - t0);
-
-    //     ops.display ();
-
-    //     // keep the first parameter that obtains > 0.5 1-recall@1
-    //     for (int i = 0; i < ops.optimal_pts.size(); i++) {
-    //         if (ops.optimal_pts[i].perf > 0.5) {
-    //             selected_params = ops.optimal_pts[i].key;
-    //             break;
-    //         }
-    //     }
-    //     assert (selected_params.size() >= 0 ||
-    //             !"could not find good enough op point");
-    // }
-
-
-    { // Use the found configuration to perform a search
-
-        knowhere::ParameterSpace params;
-
-        printf ("[%.3f s] Setting parameter configuration \"%s\" on index\n",
-                elapsed() - t0, selected_params.c_str());
-
-        params.set_index_parameters (index, selected_params.c_str());
-
-        printf ("[%.3f s] Perform a search on %ld queries\n",
-                elapsed() - t0, nq);
-
-        // output buffers
-        knowhere::SimpleIndexFlat::idx_t *I = new  knowhere::SimpleIndexFlat::idx_t[nq * k];
-        float *D = new float[nq * k];
-
-        index->search(nq, xq, k, D, I);
-
-        printf ("[%.3f s] Compute recalls\n", elapsed() - t0);
-
-        // evaluate result by hand.
-        int n_1 = 0, n_10 = 0, n_100 = 0;
-        for(int i = 0; i < nq; i++) {
-            int gt_nn = gt[i * k];
-            for(int j = 0; j < k; j++) {
-                if (I[i * k + j] == gt_nn) {
-                    if(j < 1) n_1++;
-                    if(j < 10) n_10++;
-                    if(j < 100) n_100++;
-                }
-            }
-        }
-        printf("R@1 = %.4f\n", n_1 / float(nq));
-        printf("R@10 = %.4f\n", n_10 / float(nq));
-        printf("R@100 = %.4f\n", n_100 / float(nq));
-
-        delete [] I;
-        delete [] D;
-
-    }
-
-    delete [] xq;
-    delete [] gt;
-    delete index;
-    return 0;
-}
-#endif
